+++ packages/main/src/TimePicker.ts
@@ -279,7 +279,7 @@ class TimePicker extends UI5Element implements IFormInputElement {
 	 * Defines the open or closed state of the popover.
 	 * @public
 	 * @default false
-	 * @since 2.0
+	 * @since 2.0.0
 	 */
 	@property({ type: Boolean })
 	open = false;